Abstract away statistics into comp.
